Styling
I wanted to get some of the initial styling in for the navbar which is now completed. The client has given me a website's design they wish to have for their own and so i'm working closely with the client to determine the best color theme, fonts etc.

Tomorrow i will begin on a parallax image effect for the homepage.