Parallax homepage functionality
Today i have added the functioning parallaxed hero page. I first attempted to code up the parallax from scratch and i got an effect that was very close but the issue was that the foreground images where not having the same effect and looked a little static compared to the background images, this didnt work and this code is the comments that are still here so i can see what i did later if needed. While trying to fix the issue i came to realise there was actually an npm package that would give me a customizable parallax component. Which i was able to get functional in no time. So i wasted a lot of time on my own version but i learned even more along the way so its not a loss overall, and the end result looks beautiful.