Uploading Images
Added: Admin panel - Allows the client but not the users to upload images. This is not connected yet but later the admin tab will only be visible if the logged in user is the site owner.

Added: Ability to upload images to either a hero images folder or a gallery images folder in firebase' cloud storage. This will later allow me to get the images for both the hero page and the gallery directly from this storage and render them programmatically. This will also then allow the client to simply upload , delete and swap their images with ease. This is paramount as the client is not Tech literate. So there will be a focus on abstracting complexity from the client. As this is simply a digital art gallery the actual end users will simply browse the website nothing else.

Added: Footer component - a simple component to give some links and add the copyright tag.