Functionality
I have been working yesterday and today on the bulk of this functionality and forgot to do a commit sooner, so i have added quite a bit.

firstly: I have added a gallery page which right now is simply rendering the images from an array of image URL's which will soon be replaced by the array of url's from the database.

Added: Contact page - I have added and styled a contact page which again for now is using a hard coded image but that will be changed. I have also added a fully functioning automated email service using emailJS, users can now leave their contact info with comments and an automated email will sent to the client. its capped for now at 200 a month but paid options can increase that.

Added: Admin page - The admin page allows for the uploading of images for either the homepage of just the gallery page using firebase as a back-end. This needs some work still and form validation across the board is yet to be implemented. For the hero image we are having to save it also to the firestore as well as the storage bucket so that we can have the client give their own overlay text for the image.

Refactor: I previously had several parallax elements in the hero component but i have refactored it into its own component which is much cleaner.

Added: Navbar is fully functional.

Added: Hero page - The hero page is now fully synced to the database so the images and text being used to create the hero parallax images is coming directly from firebase as will the gallery and contact images soon.

Added: Gallery card for images - I have made a custom component for the images as the client wished for all the images to be rendered as the same size and this makes it easier to dynamically render them from the database.