Gallery Firebase Sync
I have connected the gallery component to the database and i am now rendering the images from there directly. This will be great for the client as they can then simply upload their images and it will show on screen immediately.

I have also added some validation to the uploads, ensuring that it is only attempting to upload if all pieces of required information are given. I still need to implement some visuals for this on the user end so they know what is happening.

Styling changes.