Sign up and Sign in pages
Both the sign in and sign up pages have been styled and had the navbar added.

I have also made the admin panel a protected route linked to the clients uid so that only they will be shown and directed to the admin panel even if the url is known and typed in it won't show unless the logged in user is an admin. On the firebase side i have also added some rules to prevent any user with a uid not matching the owners from being able to write to the database or storage.