Delete Images
I have implemented the feature for deleting a gallery image. This became tricky as i needed the fullPath of the image that is dynamically chosen by the user (the image they want to delete) in order to delete it and i needed it where the delete button resides which is in the galleryImage component. So with a small refactor and thanks to firebase providing the full path of the image from the listAll methods results, which was incredibly helpful. Then i encountered an issue where my components where not re-rendering when an item was deleted so at the same time as deleting it from the firebase storage i am now also filtering the imageUrls state to immediately reflect the change and cause a re-render.