Image Sorting
The client has requested they be able to order the images however they would like. After some discussion we decided that it would be best if the client could input a number with the image upload to order it, i.e 1 would be the first image.

This was quite difficult to add with the implementation i had opted for when getting the gallery images. The fix i found for this was to save the image data to a seperate database, the same as the hero page. Then with some re-factoring of the logic to get the images to work with it and a very simple sorting algorithm i have achieved the desired effect.

To combat issues with the user inputting a number that already has an image with that position i am checking if there isRoom and if not then we simply slide the position of any image with a postion === newImagePosition or > newImagePosition by + 1. This makes the space needed for the new image.

I have also added the beginnings of the responsive design but i have not implemented it as of yet.