Hero Image Ordering
I have implemented the same thing from the gallery to the hero page allowing the client to choose the order of the images on their hero page, this was much easier than the gallery as we were already saving images to a seperate database with some info we simply add the position in there. Then update all the other positions accordingly as well.