Optimizations
I have done some further optimizations to the site to improve the lighthouse score, I have improved first content paint i have fixed all large layout shift issues on all pages by rendering a blank but exact size component in place of the actual one while waiting for the actual content. I still seem to be having an issue with netlify where the lighthouse SEO score suffers as a result of "robots.txt is not valid". Despite this however it still now scores 91 on SEO ive looked into this issue as i have never encountered it before. There are a couple things ive found firstly, it seems that the _redirects folder that i am using to make react router work with netlify may be part of the problem as what i can see is that the file its attempting to read as the "robots.txt" is in fact index.html, which is where my redirects lead. The other is that this issue may be because i do not have a robots.txt file. I will investigate further and try to fix it but for now the site is all green on the lighthouse report.