Debugging
Fixed - An issue where uploading a new homepage image would not trigger a re-render. I have used more of a "hack" than a solution for the gallery upload which was to simply put the user back to the homepage after a new image upload meaning once they do navigate to the gallery it will re-render. This was not going to work for the hero image as they are the homepage and moving the user away from here to force a re-render would not be good. So i have adapted the way i am accessing the images to be a live snapshot that updates the state whenever firebase' firestore updates. I have also done the same to my gallery component as i feel this is a much cleaner solution overall.