Refactor and updates
As this is moving into the production phase i have moved over to my own firebase database as opposed to the clients so that i can host my own version for my portfolio. And not have it interfere with the clients actual operations.

I also fixed an issue where my brightness was not being applied to the hero images correctly resulting in some hard to read text. This was fixed by simply tagging the line of css as important.