Deleting hero page images
I have added the functionality for the admin to see and use a delete button on the homepage images for east removal. It works in the same way as the gallery delete image function more or less. The image is first deleted from the existing state inside the component then it is deleted from the firestore database and firebase cloud storage.